+++ rs-engine/src/phases/input.rs
@@ -102,7 +102,7 @@ impl Engine {
     ///
     /// * Sets `active.player.afk_event_ready` to `true` or `false`.
     #[inline(always)]
-    fn check_afk(clock: u64, active: &mut ActivePlayer) {
+    fn check_afk(clock: u32, active: &mut ActivePlayer) {
         if clock.is_multiple_of(500) {
             let chance = if active.player.last_afk_zone == 1000 {
                 AFK_CHANCE2
